koenkk/zigbee-herdsman-converters#4257 move ubisys converters to devices file
