+++ converters/toZigbee.js
@@ -20,13 +20,6 @@ const manufacturerOptions = {
     hue: {manufacturerCode: herdsman.Zcl.ManufacturerCode.PHILIPS},
     ikea: {manufacturerCode: herdsman.Zcl.ManufacturerCode.IKEA_OF_SWEDEN},
     sinope: {manufacturerCode: herdsman.Zcl.ManufacturerCode.SINOPE_TECH},
-    /*
-     * Ubisys doesn't accept a manufacturerCode on some commands
-     * This bug has been reported, but it has not been fixed:
-     * https://github.com/Koenkk/zigbee-herdsman/issues/52
-     */
-    ubisys: {manufacturerCode: herdsman.Zcl.ManufacturerCode.UBISYS},
-    ubisysNull: {manufacturerCode: null},
     tint: {manufacturerCode: herdsman.Zcl.ManufacturerCode.MUELLER_LICHT_INT},
     legrand: {manufacturerCode: herdsman.Zcl.ManufacturerCode.VANTAGE, disableDefaultResponse: true},
     viessmann: {manufacturerCode: herdsman.Zcl.ManufacturerCode.VIESSMAN_ELEKTRO},
@@ -4393,421 +4386,6 @@ const converters = {
             } else {
                 throw new Error('LevelControl not supported on this endpoint.');
             }
-        },
-    },
-    ubisys_configure_j1: {
-        key: ['configure_j1'],
-        convertSet: async (entity, key, value, meta) => {
-            const log = (message) => {
-                meta.logger.warn(`ubisys: ${message}`);
-            };
-            const sleepSeconds = async (s) => {
-                return new Promise((resolve) => setTimeout(resolve, s * 1000));
-            };
-            const waitUntilStopped = async () => {
-                let operationalStatus = 0;
-                do {
-                    await sleepSeconds(2);
-                    operationalStatus = (await entity.read('closuresWindowCovering',
-                        ['operationalStatus'])).operationalStatus;
-                } while (operationalStatus != 0);
-                await sleepSeconds(2);
-            };
-            const writeAttrFromJson = async (attr, jsonAttr = attr, converterFunc, delaySecondsAfter) => {
-                if (jsonAttr.startsWith('ubisys')) {
-                    jsonAttr = jsonAttr.substring(6, 1).toLowerCase + jsonAttr.substring(7);
-                }
-                if (value.hasOwnProperty(jsonAttr)) {
-                    let attrValue = value[jsonAttr];
-                    if (converterFunc) {
-                        attrValue = converterFunc(attrValue);
-                    }
-                    const attributes = {};
-                    attributes[attr] = attrValue;
-                    await entity.write('closuresWindowCovering', attributes, manufacturerOptions.ubisys);
-                    if (delaySecondsAfter) {
-                        await sleepSeconds(delaySecondsAfter);
-                    }
-                }
-            };
-            const stepsPerSecond = value.steps_per_second || 50;
-            const hasCalibrate = value.hasOwnProperty('calibrate');
-            // cancel any running calibration
-            let mode = (await entity.read('closuresWindowCovering', ['windowCoveringMode'])).windowCoveringMode;
-            const modeCalibrationBitMask = 0x02;
-            if (mode & modeCalibrationBitMask) {
-                await entity.write('closuresWindowCovering', {windowCoveringMode: mode & ~modeCalibrationBitMask});
-                await sleepSeconds(2);
-            }
-            // delay a bit if reconfiguring basic configuration attributes
-            await writeAttrFromJson('windowCoveringType', undefined, undefined, 2);
-            await writeAttrFromJson('configStatus', undefined, undefined, 2);
-            if (await writeAttrFromJson('windowCoveringMode', undefined, undefined, 2)) {
-                mode = value['windowCoveringMode'];
-            }
-            if (hasCalibrate) {
-                log('Cover calibration starting...');
-                // first of all, move to top position to not confuse calibration later
-                log('  Moving cover to top position to get a good starting point...');
-                await entity.command('closuresWindowCovering', 'upOpen');
-                await waitUntilStopped();
-                log('  Settings some attributes...');
-                // reset attributes
-                await entity.write('closuresWindowCovering', {
-                    installedOpenLimitLiftCm: 0,
-                    installedClosedLimitLiftCm: 240,
-                    installedOpenLimitTiltDdegree: 0,
-                    installedClosedLimitTiltDdegree: 900,
-                    ubisysLiftToTiltTransitionSteps: 0xffff,
-                    ubisysTotalSteps: 0xffff,
-                    ubisysLiftToTiltTransitionSteps2: 0xffff,
-                    ubisysTotalSteps2: 0xffff,
-                }, manufacturerOptions.ubisys);
-                // enable calibration mode
-                await sleepSeconds(2);
-                await entity.write('closuresWindowCovering', {windowCoveringMode: mode | modeCalibrationBitMask});
-                await sleepSeconds(2);
-                // move down a bit and back up to detect upper limit
-                log('  Moving cover down a bit...');
-                await entity.command('closuresWindowCovering', 'downClose');
-                await sleepSeconds(5);
-                await entity.command('closuresWindowCovering', 'stop');
-                await sleepSeconds(2);
-                log('  Moving up again to detect upper limit...');
-                await entity.command('closuresWindowCovering', 'upOpen');
-                await waitUntilStopped();
-                log('  Moving down to count steps from open to closed...');
-                await entity.command('closuresWindowCovering', 'downClose');
-                await waitUntilStopped();
-                log('  Moving up to count steps from closed to open...');
-                await entity.command('closuresWindowCovering', 'upOpen');
-                await waitUntilStopped();
-            }
-            // now write any attribute values present in JSON
-            await writeAttrFromJson('installedOpenLimitLiftCm');
-            await writeAttrFromJson('installedClosedLimitLiftCm');
-            await writeAttrFromJson('installedOpenLimitTiltDdegree');
-            await writeAttrFromJson('installedClosedLimitTiltDdegree');
-            await writeAttrFromJson('ubisysTurnaroundGuardTime');
-            await writeAttrFromJson('ubisysLiftToTiltTransitionSteps');
-            await writeAttrFromJson('ubisysTotalSteps');
-            await writeAttrFromJson('ubisysLiftToTiltTransitionSteps2');
-            await writeAttrFromJson('ubisysTotalSteps2');
-            await writeAttrFromJson('ubisysAdditionalSteps');
-            await writeAttrFromJson('ubisysInactivePowerThreshold');
-            await writeAttrFromJson('ubisysStartupSteps');
-            // some convenience functions to not have to calculate
-            await writeAttrFromJson('ubisysTotalSteps', 'open_to_closed_s', (s) => s * stepsPerSecond);
-            await writeAttrFromJson('ubisysTotalSteps2', 'closed_to_open_s', (s) => s * stepsPerSecond);
-            await writeAttrFromJson('ubisysLiftToTiltTransitionSteps', 'lift_to_tilt_transition_ms',
-                (s) => s * stepsPerSecond / 1000);
-            await writeAttrFromJson('ubisysLiftToTiltTransitionSteps2', 'lift_to_tilt_transition_ms',
-                (s) => s * stepsPerSecond / 1000);
-            if (hasCalibrate) {
-                log('  Finalizing calibration...');
-                // disable calibration mode again
-                await sleepSeconds(2);
-                await entity.write('closuresWindowCovering', {windowCoveringMode: mode & ~modeCalibrationBitMask});
-                await sleepSeconds(2);
-                // re-read and dump all relevant attributes
-                log('  Done - will now read back the results.');
-                converters.ubisys_configure_j1.convertGet(entity, key, meta);
-            }
-        },
-        convertGet: async (entity, key, meta) => {
-            const log = (json) => {
-                meta.logger.warn(`ubisys: Cover configuration read: ${JSON.stringify(json)}`);
-            };
-            log(await entity.read('closuresWindowCovering', [
-                'windowCoveringType',
-                'physicalClosedLimitLiftCm',
-                'physicalClosedLimitTiltDdegree',
-                'installedOpenLimitLiftCm',
-                'installedClosedLimitLiftCm',
-                'installedOpenLimitTiltDdegree',
-                'installedClosedLimitTiltDdegree',
-            ]));
-            log(await entity.read('closuresWindowCovering', [
-                'configStatus',
-                'windowCoveringMode',
-                'currentPositionLiftPercentage',
-                'currentPositionLiftCm',
-                'currentPositionTiltPercentage',
-                'currentPositionTiltDdegree',
-                'operationalStatus',
-            ]));
-            log(await entity.read('closuresWindowCovering', [
-                'ubisysTurnaroundGuardTime',
-                'ubisysLiftToTiltTransitionSteps',
-                'ubisysTotalSteps',
-                'ubisysLiftToTiltTransitionSteps2',
-                'ubisysTotalSteps2',
-                'ubisysAdditionalSteps',
-                'ubisysInactivePowerThreshold',
-                'ubisysStartupSteps',
-            ], manufacturerOptions.ubisys));
-        },
-    },
-    ubisys_dimmer_setup: {
-        key: ['capabilities_forward_phase_control',
-            'capabilities_reverse_phase_control',
-            'capabilities_reactance_discriminator',
-            'capabilities_configurable_curve',
-            'capabilities_overload_detection',
-            'status_forward_phase_control',
-            'status_reverse_phase_control',
-            'status_overload',
-            'status_capacitive_load',
-            'status_inductive_load',
-            'mode_phase_control'],
-        convertSet: async (entity, key, value, meta) => {
-            if (key === 'mode_phase_control') {
-                const phaseControl = value.toLowerCase();
-                const phaseControlValues = {'automatic': 0, 'forward': 1, 'reverse': 2};
-                utils.validateValue(phaseControl, Object.keys(phaseControlValues));
-                await entity.write('manuSpecificUbisysDimmerSetup',
-                    {'mode': phaseControlValues[phaseControl]}, manufacturerOptions.ubisysNull);
-            }
-            converters.ubisys_dimmer_setup.convertGet(entity, key, meta);
-        },
-        convertGet: async (entity, key, meta) => {
-            await entity.read('manuSpecificUbisysDimmerSetup', ['capabilities'], manufacturerOptions.ubisysNull);
-            await entity.read('manuSpecificUbisysDimmerSetup', ['status'], manufacturerOptions.ubisysNull);
-            await entity.read('manuSpecificUbisysDimmerSetup', ['mode'], manufacturerOptions.ubisysNull);
-        },
-    },
-    ubisys_dimmer_setup_genLevelCtrl: {
-        key: ['minimum_on_level'],
-        convertSet: async (entity, key, value, meta) => {
-            if (key === 'minimum_on_level') {
-                await entity.write('genLevelCtrl', {'ubisysMinimumOnLevel': value}, manufacturerOptions.ubisys);
-            }
-            converters.ubisys_dimmer_setup_genLevelCtrl.convertGet(entity, key, meta);
-        },
-        convertGet: async (entity, key, meta) => {
-            await entity.read('genLevelCtrl', ['ubisysMinimumOnLevel'], manufacturerOptions.ubisys);
-        },
-    },
-    ubisys_device_setup: {
-        key: ['configure_device_setup'],
-        convertSet: async (entity, key, value, meta) => {
-            const devMgmtEp = meta.device.getEndpoint(232);
-
-            if (value.hasOwnProperty('input_configurations')) {
-                // example: [0, 0, 0, 0]
-                await devMgmtEp.write(
-                    'manuSpecificUbisysDeviceSetup',
-                    {'inputConfigurations': {elementType: 'data8', elements: value.input_configurations}},
-                    manufacturerOptions.ubisysNull,
-                );
-            }
-
-            if (value.hasOwnProperty('input_actions')) {
-                // example (default for C4): [[0,13,1,6,0,2], [1,13,2,6,0,2], [2,13,3,6,0,2], [3,13,4,6,0,2]]
-                await devMgmtEp.write(
-                    'manuSpecificUbisysDeviceSetup',
-                    {'inputActions': {elementType: 'octetStr', elements: value.input_actions}},
-                    manufacturerOptions.ubisysNull,
-                );
-            }
-
-            if (value.hasOwnProperty('input_action_templates')) {
-                const templateTypes = {
-                    // source: "ZigBee Device Physical Input Configurations Integrator’s Guide"
-                    // (can be obtained directly from ubisys upon request)
-                    'toggle': {
-                        getInputActions: (input, endpoint) => [
-                            [input, 0x0D, endpoint, 0x06, 0x00, 0x02],
-                        ],
-                    },
-                    'toggle_switch': {
-                        getInputActions: (input, endpoint) => [
-                            [input, 0x0D, endpoint, 0x06, 0x00, 0x02],
-                            [input, 0x03, endpoint, 0x06, 0x00, 0x02],
-                        ],
-                    },
-                    'on_off_switch': {
-                        getInputActions: (input, endpoint) => [
-                            [input, 0x0D, endpoint, 0x06, 0x00, 0x01],
-                            [input, 0x03, endpoint, 0x06, 0x00, 0x00],
-                        ],
-                    },
-                    'on': {
-                        getInputActions: (input, endpoint) => [
-                            [input, 0x0D, endpoint, 0x06, 0x00, 0x01],
-                        ],
-                    },
-                    'off': {
-                        getInputActions: (input, endpoint) => [
-                            [input, 0x0D, endpoint, 0x06, 0x00, 0x00],
-                        ],
-                    },
-                    'dimmer_single': {
-                        getInputActions: (input, endpoint, template) => {
-                            const moveUpCmd = template.no_onoff || template.no_onoff_up ? 0x01 : 0x05;
-                            const moveDownCmd = template.no_onoff || template.no_onoff_down ? 0x01 : 0x05;
-                            const moveRate = template.rate || 50;
-                            return [
-                                [input, 0x07, endpoint, 0x06, 0x00, 0x02],
-                                [input, 0x86, endpoint, 0x08, 0x00, moveUpCmd, 0x00, moveRate],
-                                [input, 0xC6, endpoint, 0x08, 0x00, moveDownCmd, 0x01, moveRate],
-                                [input, 0x0B, endpoint, 0x08, 0x00, 0x03],
-                            ];
-                        },
-                    },
-                    'dimmer_double': {
-                        doubleInputs: true,
-                        getInputActions: (inputs, endpoint, template) => {
-                            const moveUpCmd = template.no_onoff || template.no_onoff_up ? 0x01 : 0x05;
-                            const moveDownCmd = template.no_onoff || template.no_onoff_down ? 0x01 : 0x05;
-                            const moveRate = template.rate || 50;
-                            return [
-                                [inputs[0], 0x07, endpoint, 0x06, 0x00, 0x01],
-                                [inputs[0], 0x06, endpoint, 0x08, 0x00, moveUpCmd, 0x00, moveRate],
-                                [inputs[0], 0x0B, endpoint, 0x08, 0x00, 0x03],
-                                [inputs[1], 0x07, endpoint, 0x06, 0x00, 0x00],
-                                [inputs[1], 0x06, endpoint, 0x08, 0x00, moveDownCmd, 0x01, moveRate],
-                                [inputs[1], 0x0B, endpoint, 0x08, 0x00, 0x03],
-                            ];
-                        },
-                    },
-                    'cover': {
-                        cover: true,
-                        doubleInputs: true,
-                        getInputActions: (inputs, endpoint) => [
-                            [inputs[0], 0x0D, endpoint, 0x02, 0x01, 0x00],
-                            [inputs[0], 0x07, endpoint, 0x02, 0x01, 0x02],
-                            [inputs[1], 0x0D, endpoint, 0x02, 0x01, 0x01],
-                            [inputs[1], 0x07, endpoint, 0x02, 0x01, 0x02],
-                        ],
-                    },
-                    'cover_switch': {
-                        cover: true,
-                        doubleInputs: true,
-                        getInputActions: (inputs, endpoint) => [
-                            [inputs[0], 0x0D, endpoint, 0x02, 0x01, 0x00],
-                            [inputs[0], 0x03, endpoint, 0x02, 0x01, 0x02],
-                            [inputs[1], 0x0D, endpoint, 0x02, 0x01, 0x01],
-                            [inputs[1], 0x03, endpoint, 0x02, 0x01, 0x02],
-                        ],
-                    },
-                    'cover_up': {
-                        cover: true,
-                        getInputActions: (input, endpoint) => [
-                            [input, 0x0D, endpoint, 0x02, 0x01, 0x00],
-                        ],
-                    },
-                    'cover_down': {
-                        cover: true,
-                        getInputActions: (input, endpoint) => [
-                            [input, 0x0D, endpoint, 0x02, 0x01, 0x01],
-                        ],
-                    },
-                    'scene': {
-                        scene: true,
-                        getInputActions: (input, endpoint, groupId, sceneId) => [
-                            [input, 0x07, endpoint, 0x05, 0x00, 0x05, groupId & 0xff, groupId >> 8, sceneId],
-                        ],
-                        getInputActions2: (input, endpoint, groupId, sceneId) => [
-                            [input, 0x06, endpoint, 0x05, 0x00, 0x05, groupId & 0xff, groupId >> 8, sceneId],
-                        ],
-                    },
-                    'scene_switch': {
-                        scene: true,
-                        getInputActions: (input, endpoint, groupId, sceneId) => [
-                            [input, 0x0D, endpoint, 0x05, 0x00, 0x05, groupId & 0xff, groupId >> 8, sceneId],
-                        ],
-                        getInputActions2: (input, endpoint, groupId, sceneId) => [
-                            [input, 0x03, endpoint, 0x05, 0x00, 0x05, groupId & 0xff, groupId >> 8, sceneId],
-                        ],
-                    },
-                };
-
-                // first input
-                let input = 0;
-                // first client endpoint - depends on actual device
-                let endpoint = {'S1': 2, 'S2': 3, 'D1': 2, 'J1': 2, 'C4': 1}[meta.mapped.model];
-                // default group id
-                let groupId = 0;
-
-                const templates = Array.isArray(value.input_action_templates) ? value.input_action_templates :
-                    [value.input_action_templates];
-                let resultingInputActions = [];
-                for (const template of templates) {
-                    const templateType = templateTypes[template.type];
-                    if (!templateType) {
-                        throw new Error(`input_action_templates: Template type '${template.type}' is not valid ` +
-                            `(valid types: ${Object.keys(templateTypes)})`);
-                    }
-
-                    if (template.hasOwnProperty('input')) {
-                        input = template.input;
-                    }
-                    if (template.hasOwnProperty('endpoint')) {
-                        endpoint = template.endpoint;
-                    }
-                    // C4 cover endpoints only start at 5
-                    if (templateType.cover && meta.mapped.model === 'C4' && endpoint < 5) {
-                        endpoint += 4;
-                    }
-
-                    let inputActions;
-                    if (!templateType.doubleInputs) {
-                        if (!templateType.scene) {
-                            // single input, no scene(s)
-                            inputActions = templateType.getInputActions(input, endpoint, template);
-                        } else {
-                            // scene(s) (always single input)
-                            if (!template.hasOwnProperty('scene_id')) {
-                                throw new Error(`input_action_templates: Need an attribute 'scene_id' for '${template.type}'`);
-                            }
-                            if (template.hasOwnProperty('group_id')) {
-                                groupId = template.group_id;
-                            }
-                            inputActions = templateType.getInputActions(input, endpoint, groupId, template.scene_id);
-
-                            if (template.hasOwnProperty('scene_id_2')) {
-                                if (template.hasOwnProperty('group_id_2')) {
-                                    groupId = template.group_id_2;
-                                }
-                                inputActions = inputActions.concat(templateType.getInputActions2(input, endpoint, groupId,
-                                    template.scene_id_2));
-                            }
-                        }
-                    } else {
-                        // double inputs
-                        input = template.hasOwnProperty('inputs') ? template.inputs : [input, input + 1];
-                        inputActions = templateType.getInputActions(input, endpoint, template);
-                    }
-                    resultingInputActions = resultingInputActions.concat(inputActions);
-
-                    meta.logger.warn(`ubisys: Using input(s) ${input} and endpoint ${endpoint} for '${template.type}'.`);
-                    // input might by now be an array (in case of double inputs)
-                    input = (Array.isArray(input) ? Math.max(...input) : input) + 1;
-                    endpoint += 1;
-                }
-
-                meta.logger.debug(`ubisys: input_actions to be sent to '${meta.options.friendlyName}': ` +
-                    JSON.stringify(resultingInputActions));
-                await devMgmtEp.write(
-                    'manuSpecificUbisysDeviceSetup',
-                    {'inputActions': {elementType: 'octetStr', elements: resultingInputActions}},
-                    manufacturerOptions.ubisysNull,
-                );
-            }
-
-            // re-read effective settings and dump them to the log
-            converters.ubisys_device_setup.convertGet(entity, key, meta);
-        },
-
-        convertGet: async (entity, key, meta) => {
-            const log = (dataRead) => {
-                meta.logger.warn(
-                    `ubisys: Device setup read for '${meta.options.friendly_name}': ${JSON.stringify(utils.toSnakeCase(dataRead))}`);
-            };
-            const devMgmtEp = meta.device.getEndpoint(232);
-            log(await devMgmtEp.read('manuSpecificUbisysDeviceSetup', ['inputConfigurations'], manufacturerOptions.ubisysNull));
-            log(await devMgmtEp.read('manuSpecificUbisysDeviceSetup', ['inputActions'], manufacturerOptions.ubisysNull));
         },
     },
     tint_scene: {
